+++ python/ray/serve/tests/test_metrics.py
@@ -13,15 +13,15 @@ from ray import serve
 from ray._private.test_utils import SignalActor, wait_for_condition
 from ray.serve._private.constants import DEFAULT_LATENCY_BUCKET_MS
 from ray.serve._private.long_poll import LongPollHost, UpdatedObject
-from ray.serve._private.utils import block_until_http_ready
-from ray.serve.config import gRPCOptions
-from ray.serve.handle import DeploymentHandle
-from ray.serve.metrics import Counter, Gauge, Histogram
-from ray.serve.tests.common.utils import (
+from ray.serve._private.test_utils import (
     ping_fruit_stand,
     ping_grpc_call_method,
     ping_grpc_list_applications,
 )
+from ray.serve._private.utils import block_until_http_ready
+from ray.serve.config import gRPCOptions
+from ray.serve.handle import DeploymentHandle
+from ray.serve.metrics import Counter, Gauge, Histogram
 from ray.serve.tests.test_config_files.grpc_deployment import g, g2
