+++ python/ray/serve/tests/unit/test_application_state.py
@@ -20,10 +20,10 @@ from ray.serve._private.common import (
 from ray.serve._private.config import DeploymentConfig, ReplicaConfig
 from ray.serve._private.deploy_utils import deploy_args_to_deployment_info
 from ray.serve._private.deployment_info import DeploymentInfo
+from ray.serve._private.test_utils import MockKVStore
 from ray.serve._private.utils import get_random_letters
 from ray.serve.exceptions import RayServeException
 from ray.serve.schema import DeploymentSchema, ServeApplicationSchema
-from ray.serve.tests.common.utils import MockKVStore
 
 
 class MockEndpointState:
