+++ python/ray/serve/tests/unit/test_proxy_request_response.py
@@ -9,8 +9,8 @@ from ray.serve._private.proxy_request_response import (
     ProxyRequest,
     gRPCProxyRequest,
 )
+from ray.serve._private.test_utils import FakeGrpcContext
 from ray.serve.generated import serve_pb2
-from ray.serve.tests.common.utils import FakeGrpcContext
 
 
 class TestASGIProxyRequest:
